CAUSEWAY-2297: ExcelFileWriter: removes row limit 32767

+++ extensions/vw/tabular/excel/src/main/java/org/apache/causeway/extensions/tabular/excel/exporter/ExcelFileWriter.java
@@ -113,7 +113,7 @@ public record ExcelFileWriter(@Nullable Options options) {
         private final Sheet sheet;
         private int rowNum;
         public Row newRow() {
-            return sheet.createRow((short) rowNum++);
+            return sheet.createRow(rowNum++);
         }
     }
